refactor: addresses
