+++ src/routes/addresses/address/extended.ts
@@ -1,0 +1,118 @@
+import { getOnchainMetadata, getSchemaForEndpoint } from '@blockfrost/openapi';
+import { FastifyInstance, FastifyRequest } from 'fastify';
+
+import { SQLQuery } from '../../../sql';
+import * as QueryTypes from '../../../types/queries/addresses';
+import * as ResponseTypes from '../../../types/responses/addresses';
+import { getDbSync } from '../../../utils/database';
+import { handle404, handleInvalidAddress } from '../../../utils/error-handler';
+import { fetchAssetMetadata } from '../../../utils/token-registry';
+import {
+  getAddressTypeAndPaymentCred,
+  paymentCredToBech32Address,
+} from '../../../utils/validation';
+
+async function route(fastify: FastifyInstance) {
+  fastify.route({
+    url: '/addresses/:address/extended',
+    method: 'GET',
+    schema: getSchemaForEndpoint('/addresses/{address}/extended'),
+    handler: async (request: FastifyRequest<QueryTypes.RequestParameters>, reply) => {
+      const { addressType, paymentCred } = getAddressTypeAndPaymentCred(request.params.address);
+
+      if (!addressType) {
+        return handleInvalidAddress(reply);
+      }
+
+      const clientDbSync = await getDbSync(fastify);
+
+      try {
+        const query404 = await clientDbSync.query<QueryTypes.ResultFound>(
+          SQLQuery.get('addresses_404'),
+          [request.params.address, paymentCred],
+        );
+
+        if (query404.rows.length === 0) {
+          clientDbSync.release();
+          return handle404(reply);
+        }
+        const { rows } = await clientDbSync.query<QueryTypes.AddressExtendedQuery>(
+          SQLQuery.get('addresses_address_extended'),
+          [request.params.address, paymentCred],
+        );
+
+        clientDbSync.release();
+
+        // if paymentCred is used we have to convert it back to bech32
+        if (paymentCred) {
+          const bech32paymentCred = paymentCredToBech32Address(rows[0].address);
+
+          if (bech32paymentCred) rows[0].address = bech32paymentCred;
+        }
+
+        const assetsAmount: ResponseTypes.AmountExtended = [];
+
+        // add off-chain data to all assets if they exist
+        if (rows[0].amount) {
+          for (const asset of rows[0].amount) {
+            const unit = `${asset.policy_id}${asset.asset_name}`;
+            const registryData = await fetchAssetMetadata(unit);
+            const { onchainMetadata } = getOnchainMetadata(
+              asset.onchain_metadata,
+              asset.asset_name,
+              asset.policy_id,
+            );
+
+            assetsAmount.push({
+              unit: unit,
+              quantity: asset.quantity,
+              decimals: registryData?.decimals ?? null,
+              has_nft_onchain_metadata: onchainMetadata !== null,
+            });
+          }
+        }
+
+        // quantities/amounts are returned as string from database so they won't overflow JS number
+        const result: ResponseTypes.AddressExtended = rows[0].amount
+          ? {
+              address: rows[0].address,
+              amount: [
+                {
+                  unit: 'lovelace',
+                  quantity: rows[0].amount_lovelace,
+                  decimals: 6,
+                  has_nft_onchain_metadata: false,
+                },
+                ...assetsAmount,
+              ],
+              stake_address: rows[0].stake_address,
+              type: addressType,
+              script: rows[0].script,
+            }
+          : {
+              address: rows[0].address,
+              amount: [
+                {
+                  unit: 'lovelace',
+                  quantity: rows[0].amount_lovelace,
+                  decimals: 6,
+                  has_nft_onchain_metadata: false,
+                },
+              ],
+              stake_address: rows[0].stake_address,
+              type: addressType,
+              script: rows[0].script,
+            };
+
+        return reply.send(result);
+      } catch (error) {
+        if (clientDbSync) {
+          clientDbSync.release();
+        }
+        throw error;
+      }
+    },
+  });
+}
+
+export default route;
